+++ llama_index/vector_stores/weaviate_utils.py
@@ -6,6 +6,7 @@ Contain conversion to and from dataclasses that LlamaIndex uses.
 
 import json
 import logging
+import math
 from dataclasses import field
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, cast
 
@@ -131,6 +132,12 @@ def _legacy_metadata_dict_to_node(entry: Dict[str, Any]) -> Tuple[dict, dict, di
     return extra_info, node_info, relationships
 
 
+def get_node_similarity(entry: Dict, similarity_key: str = "distance") -> float:
+    """Get converted node similarity from distance."""
+    distance = float(entry["_additional"][similarity_key])
+    return 1.0 - math.exp(-distance)
+
+
 def to_node(entry: Dict, text_key: str = DEFAULT_TEXT_KEY) -> Node:
     """Convert to Node."""
     additional = entry.pop("_additional")
@@ -148,6 +155,7 @@ def to_node(entry: Dict, text_key: str = DEFAULT_TEXT_KEY) -> Node:
         extra_info=extra_info,
         node_info=node_info,
         relationships=relationships,
+        embedding=additional["vector"],
     )